# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Log in with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200':
          description: Logged in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser'
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: Get the current authenticated user
      responses:
        '200':
          description: Current user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /auth/reset-password:
    post:
      tags: [Auth]
      operationId: resetPassword
      summary: Reset a staff password using the recovery code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordInput'
      responses:
        '200':
          description: Password updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Invalid recovery code or username.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: Password reset is not configured on this server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /auth/setup:
    post:
      tags: [Auth]
      operationId: setupAdmin
      summary: Create the first staff account when none exist yet
      description: >
        Only succeeds when the staff_users table is empty. Creates the account
        and signs the caller in. After the first account exists, this endpoint
        returns 409.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetupAdminInput'
      responses:
        '200':
          description: First admin created and signed in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: An admin account already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
...
    SetupAdminInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 6
    ResetPasswordInput:
      type: object
      required: [username, resetCode, newPassword]
      properties:
        username:
          type: string
          minLength: 1
        resetCode:
          type: string
          minLength: 1
          description: Recovery code configured as PASSWORD_RESET_SECRET on the server.
        newPassword:
          type: string
          minLength: 6
    AuthUser:
      type: object
      required: [id, username]
      properties:
        id:
          type: integer
        username:
          type: string