# lib/api-zod/src/generated/types/setupAdminInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SetupAdminInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 6 */
  password: string;
}